Add docs for sending
