+++ src/main/java/net/dv8tion/jda/internal/entities/mixin/channel/middleman/GuildMessageChannelMixin.java
@@ -19,11 +19,14 @@ package net.dv8tion.jda.internal.entities.mixin.channel.middleman;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.GuildMessageChannel;
 import net.dv8tion.jda.api.entities.User;
+import net.dv8tion.jda.api.entities.sticker.StickerSnowflake;
 import net.dv8tion.jda.api.requests.RestAction;
+import net.dv8tion.jda.api.requests.restaction.MessageAction;
 import net.dv8tion.jda.api.utils.MiscUtil;
 import net.dv8tion.jda.api.utils.TimeUtil;
 import net.dv8tion.jda.internal.requests.RestActionImpl;
 import net.dv8tion.jda.internal.requests.Route;
+import net.dv8tion.jda.internal.requests.restaction.MessageActionImpl;
 import net.dv8tion.jda.internal.utils.Checks;
 import net.dv8tion.jda.internal.utils.EncodingUtil;
 
@@ -100,7 +103,16 @@ public interface GuildMessageChannelMixin<T extends GuildMessageChannelMixin<T>>
         Route.CompiledRoute route = Route.Messages.CLEAR_EMOTE_REACTIONS.compile(getId(), messageId, code);
         return new RestActionImpl<>(getJDA(), route);
     }
-    
+
+    @Nonnull
+    @Override
+    default MessageAction sendStickers(@Nonnull Collection<? extends StickerSnowflake> stickers)
+    {
+        checkCanAccessChannel();
+        checkCanSendMessage();
+        return new MessageActionImpl(getJDA(), null, this).setStickers(stickers);
+    }
+
     // ---- Default implementation of parent mixins hooks ----
     default void checkCanAccessChannel()
     {
